Split comment form out to its’ own class
